Switch to using pages and react-router-dom
- Move realtime test and motion test to new pages
- Existing VoiceAssistant moved to default page
- Extracted AssistantWithOptionalIntegrations into component
