+++ packages/frontend/package.json
@@ -41,6 +41,7 @@
     "react-katex": "^3.0.1",
     "react-loader-spinner": "^6.1.6",
     "react-markdown": "^9.0.1",
+    "react-router-dom": "^6.26.2",
     "react-syntax-highlighter": "^15.5.0",
     "react-transition-group": "^4.4.5",
     "sentence-splitter": "^5.0.0",
