+++ packages/frontend/src/main.tsx
@@ -1,10 +1,10 @@
 //import React from "react"
 import ReactDOM from "react-dom/client";
-import App from "./App";
+import Root from "./Root";
 import "./index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   //  <React.StrictMode>
-  <App />,
+  <Root />,
   //  </React.StrictMode>,
 );
